requirements.txt
asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.10.0
flake8==3.8.4
gunicorn==20.0.4
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
sqlparse==0.4.1
djangorestframework~=3.12.2